In property editor: a completion selection should stop editing.
Completions in the editor inserts the selected completion as
text and then lets the user continue to edit the document.
In the property editor (inspector / all properties table / and
slice editor) we should just accept the value and stop the
editing of the property value. Since the value is always a
complete value - not just a part of the value.

Test: Added UI test for the added logic.

Change-Id: I16143c6042b3e0dee69f70191eecc9782cfb1fd6