Remove SpeedSearch from Palette.
It was confusing to have a SpeedSearch in addition to the search
box on the palette. This CL removes the SpeedSearch but will start
a search if the user types a character while the component tree
grid has focus.

Also: Reactivate AttachedToolWindowTest that were being ignored
after it started failing.

Bug: http://b.android.com/235259 http://b.android.com/235080
Test: Updated
Change-Id: I36107482ef449d8ad730fb4fe27cd8d73b9c5dc8