Removing invokeLater from StudioProgressRunner
As a convenience, StudioProgressRunner could invokeLater for you,
which was useful when running callbacks after SDK load was complete.
However, since the SDK mechanism uses the progress runner for internal
tasks (e.g. downloading things) as well, it was never (in the context of
the sdk, which I believe is the only context in which
StudioProgressRunner is used) the right thing to do to have it invoke
for you. So removing that functionality (and adding it in the callbacks
where it was done automatically before).

Fixes b.android.com/252417

Test: removed tests for this functionality from StudioProgressRunnerTest
Change-Id: I1dede0a1f1da7a0f39eea1719f7de5e0c1458c2c