New Error model and Error Panel
The CL gather the RenderingError and AndroidLintError in the same
panel.

To ensure a better consitency between for the handling of the error in
Nele, the NlIssue class has been created and both kind of error have
been wrapped into an NlIssue class.

This also provide an easier way to display any kind of message in the
IssuePanel if we need to in the future (by subclassing NlIssue and
adding Generigy IssueModel)

Test: Unit test and UI test have been written. DesignSurfaceFixture have
been modifier to use the new IssuePanel. (90% Line Coverage)

Change-Id: I0d776efcd4e01e44f4a1810c61f6682219703dac