apk analyzer: don't lock open apk files
We used to rely on IntelliJ's mechanism for copying apks/jars on
Windows in order to not lock the APK file open in the editor.

But the APK Analyzer now uses Java's file system instead of
VirtualFileSystem, so we need to implement this ourself. This new
mechanims always makes a copy of the APK, and has the editor open
the temporary copy. Whenever a change is detected to the original
file, the contents of the editor are again refreshed.

Bug: http://b.android.com/226714
Test: existing manual test - open a file in apk analyzer, and attempt
to update it by doing another build.

Change-Id: I014a9570a867d753e5e48b749d237cafba23e49a