Unify multiple "Android SDK missing" error dialogs
When the Android SDK is missing, there are 5 locations that test and
display a similar error message. We are about to add some more, for the
new Project Wizard work.
Having a central location to display this error message, will also make
it easier to offer a link to a solution in the future.

Test: Rename/delete your Android SDK location and create a new Project
Change-Id: I4d5ba55d9400eddda83f3a6358bbc618a65fda3d