Fix flaky LayoutInspectorTest.launchLayoutInspectorViaChooser
There are two sources of flakiness in the test. The first is that
the test may fail to click the menu item if that item is not shown
yet. The CL use isMenuPathEnabled check to handle it. The CL fixes a bug
of isMenuPathEnabled that if the expanded menu is more than one layer,
the entire menu would not be closed by clicking ESC only once.

The second issue is that the process the test needs to select may not be
ready in 10 seconds because during the wait, the test may still wait for
the device come online which could take up to 120 seconds (note that
guiTest.waitForBackgroundTasks() does not really wait for the device
come online). So, we change the timeout to 120 seconds which is a common
number we use to wait for a device come online.

Test: run the test.
Change-Id: I32afbdaa35bef6c82abcfe27b4d1f0c765a14332