Update base split handling to cope with nested modules
Finding the base split was being done by name which
didn't work when the base lib and split have the same
name but are in different folders. This CL updates the
logic to use the gradle path instead and adds a new test
project to test both the flat and nested cases for
multi-atom projects.

Also updated the temporary feature-module simulation code
and added a comment to make sure it gets removed when
feature-modules are added

Fixes: http://b/36668312

Test: InstantAppsTest and MultiAtomSupportTest
Change-Id: Idbb3138cebfb0a6122a704c1d0bec73c09505b86