[AIA] Skips provision task when not AIA module
After this CL, provision is skipped if the run configuration has a non
Instant App module selected.

Since when the user first creates a run configuration there is no module
selected, we can't know if it will be an AIA run config or not, so have
to create the provision task. The quick fix for the bug is just skip the
provision if not AIA.

Test: JUnit tests

Bug: http://b/36800719
Change-Id: I2de3e8f453c09b8e11124dd9f2c7651aa4d56d66