[AIA] Fix running problem in N emulators
Emulators running N API 24 or 25 don't have "test-keys" os build type,
and then we have to check the hardware to verify if it's an emulator.

This CL also changes provision to set the required flags even when the
package is already installed so we guaranteee that everything is correct
even if the user installed the packages manually.

It also swaps the order of checking if the user is signed in a google
account and check the API level, since Google account is not required
for post O devices.

Test: JUnit tests
Change-Id: If65fc91fbd323cb91f2ac369989be290c54d6796