Adding nav editor metadata and inspectors
The user-visible effect of this change is to add property inspectors for
nav destinations and actions (right panel).
Unfortunately, the properties system requires the xml metadata be
defined, so this change also includes that, giving us code completion
for nav files.
Since we needed the attributes metadata, I also switched to referencing
the actual annotated classes in the nav runtime library to find the tag
names and class hierarchy.

Test: NavigationSchemaTest
Change-Id: Iecab7acae6f8a417ab84cae3cbfa78f27de7fe36