Fixed bug where IDE tried to update the plugin version when constraint layout is missing.
This bug showed up when the NPW stopped downloading constraint
layout during project creation. Now Gradle reports that constraint
layout is missing, and the IDE thinks that this is an old version
of the plugin that do not support having constraint layout in the
SDK manager.

The fix involves changing the assumption by realizing that a
null model means that the project is a new project, that has not
been set up yet, instead of an old project that do not support
constraint layout.

Fixes
https://code.google.com/p/android/issues/detail?id=360563

Test: JUnit test.
Change-Id: I1325e3ce71d51269abb2975e42b7286f13295254