Fix kotlin anonymous code navigation issue
In at least one case with Kotlin code, our code navigation failed to
work. It was particularly a shame because it was clear that there was
more than enough information in the callstack (outer class name + line
number), but we were failing to use it, even though the underlying call
to IntelliJ's Psi code shouldn't have been returning null.

This CL makes the logic a bit more robust in the face of what is likely
edge-case Psi errors.

Bug: https://b.corp.google.com/issues/37501493
Test: Tested as much as possible. Would like to have tested IntelliJ
logic also, but their code was not designed to be tested.

Change-Id: Iaf2bef30c4404269bc9f10fce077848dac3fc275