Add dependencies to base feature
This CL works around an issue in the build
system that forces us to put all dependencies in
the base feature of an instant apps project

To do this the dependency action in the template
recipe system has been altered to update the
dependencies in the base feature if required
(and leave them out if no appropriate location
is provided).

Also the New Model Wizard has been updated to
try to find an existing base module in the project.
If none can be found then the new module itself
is added as a base module.

Fixes: http://b/37729817

Test: instantapp UI tests
Change-Id: I07bcdc9fa7c0602471af7dae696a93aba1b451f9