Refactor PTable into adt-ui
PTable is a JBPTable component that renders grouped elements. It's
currently used in designer for the properties panel. We want to reuse it
for Layout Inspector's properties panel. It has the advantage of also
sharing look and feel between two similar tools in studio.

Refactors as much of the non-NL editor item into adt-ui. Certain
features of PTable (like starring items depends on AndroidIcons, or
deleting group items or editingCanceled method) have android or intellij
dependencies that are kept in designer through the NlPTable class so
it stays in designer.

Test: refactored existing tests
Change-Id: Ibb1f27c1d1304c374a006163f7f335aea12d9ede