Use debugger UI button instead of menu to resume.
The menu way is flaky.  We could try to fix it, but I believe the
flakiness is due to interaction with the dual debugger in a way that may
mean we need to remove the menu option.

Most importantly, users are unlikely to click Run -> Resume Program, the
resume button right beside the variables and frames is way more
convenient.

Test: Improves flakiness, ran several times locally to check.
Change-Id: Iae742fd9e4269d9f223b00f506133602197d8c66