Add an ExtensionPoint for converting to kotlin
Add ConvertJavaToKotlinProvider that has methods to
a. convert java files to kotlin
b. configure kotlin in a project
c. Add a default reflection implementation that makes a
   few assumptions and has few issues with the transformed
   kotlin code.
d. Use KOTLIN_ENABLED flag to decide whether to convert the files

Also add the relevant code to call the converter as the last
thing in the render pipeline.

Lastly add the Include Kotlin Support checkbox to the
"Configure Android Project" form.

Test: None yet
Change-Id: I687240b2a6bcaa71172b9ab991f5bef73822fb08