[APK Debugging] Fixed StackOverflowError in 'Android' view.
The issue was that nodes in the tree view are used as keys
in a HashMap and for this to work properly, the nodes need
to implement equals/hashCode. Nodes by default delegate
this check to the values inside them, and ApkClass and
ApkPackage were not implementing them.

This CL fixes the StackOverflowError by implementing
equals/hashCode in ApkClass and ApkPackage.

There are a few other fixes for the 'Android' view that
are hard to separate in another CL (this CL was originally
fixing those minor issues, and then I encountered the
StackOverflowError.)

Bug: 37854131

Test: JUnit tests.
Change-Id: I2d406d1f722ff2e34961dd805165e5d27fa1fad5