Fix deactivate logic in NlPreviewForm
The previous logic to disable rendering in a deactivated preview
form could cause the form to clear the file that was supposed to
be shown when the form eventually got activated.

The field myInactiveFile got cleared because deactivate() was
called several times. This field is now removed and we use the
isActive field to identify if rendering should be blocked.

Bug: http://b/37856821

Test: The fix of the bug required manual test (see the bug).

Change-Id: I67a7cadbf4677bbf5ca7adc077ebb19691bf14b5