Fixing crash when stopping trace with unplugged phone.
Start an app
Start a CPU trace
Unplug phone
Before UI updates stop CPU trace.

This used to crash due to BufferString failing with
a null pointer exception. I have changed the function
around to support a null return. This will also
work if a request for an ID doesn't exist as such
I created a test that test for invalid id.

Test: CpuTableTest
Bug: 37955171
Change-Id: If98d100da901c104d815b30048862ba944d2b033