Now the "plugin upgrade dialog" can remind users about upgrade after 1 day.
The button "remind me later" in the "plugin upgrade dialog" simply
closed the dialog and nothing else. The "later" in "remind me later"
was really "remind me on the next sync", which is pretty annoying
to our users.

Since we don't have a notion of a "project" session, the best we
can do now is to change "remind me later" to "remind me tomorrow".
This CL implements that functionality.

Bug: 37119900

Test: JUnit tests.
Change-Id: Id328c2a410fe3bbfd42345270ef721bf287e1378