[APK Debugging] Now we use TreeVisitor to find nodes in the 'Android' view.
This CL improves node search in the 'Android' view by
using a TreeVisitor in the view's TreeStructure, instead of
relying on the actual UI nodes. The advantage of a
TreeVisitor is that nodes are guaranteed to be fully
constructed for the search, while UI nodes may not be
constructed or they could be partially constructed, resulting
in a failed search.

Besides this, this CL also:
- Adds a way to find a select a node in the 'Android' view
- Adding a DebuggableSharedObjectFile to a NativeLibrary no longer
  clears the source paths and path mappings. This was done in the
  past when a library had only one debuggable shared object library
  (.so). This is no longer the case. A library can have 1 .so file
  per ABI.
- Minor method renaming improvements in DebuggableSharedObjectFile.

Test: JUnit tests.
Change-Id: I5f2186e451fddc6f9a493b9c15c7dbcca988b79a