Studio-side database storage of method and stack
Added two additional tables: Memory_MethodInfos + Memory_StackInfos,
which can be used to resolve any encoded callstacks that are received
from an app's agent. Note that currently, an asssumption is made such
that the method and stack infos are received before or along with the
BatchAllocationSample containing allocation events that use those data.
This assumption is guaranteed by how the agent processes and sends alloc
data in the order they occur in the app.

Bug: 37175866
Test: Additional unit tests for MemoryTable
Change-Id: I9a0f8d8c699c34d2fe2c34f714bdc159832c8bd5