Set myLogger in LayoutlibCallbackImpl
If the logger is not null in the constructor, there is no need to wait
for a setLogger call.
This will avoid a crash when calling RenderTask#renderDrawable when
inflate() has not been called before.

Test: RenderTaskTest should cover this but the test renderDrawable
currently has workaround for a problem in layout. Once the new layoutlib
is in, the workaround will be removed and this tested.

Change-Id: If0d141229add3433e56c536fbadc1dfa8d7c5cec