[Modularize] Group resource items by URLs
In order to simplify the selection, we now group all associated
resource items under a "@type/name" node. Individual resource items
can no longer be selected/de-selected, because we want to either
move the resource as a whole or leave it in place.

I've included a minor cleanup for loading icons, although it's still
the case we don't have a way to listen for when icons have been
loaded and re-paint.

Test: N/A
Change-Id: I8ebed5da980644cf9d5565e974345ef9de1487be