38073284: IDE needs to check maven.google.com for new libraries
This CL hooks up the IDE side of the fix for 38073284: it instantiates
the GoogleMavenRepository on the IDE side (which lives in tools/base and
is tested there, but is subclassed here in the IDE to connect the
cache directory to the IDE cache, and to connect the network access
to the IDE's own HTTP connection factory (to ensure that it uses the
right proxy settings).

Similarly it updates the LintIdeClient to the new signatures of
the caching and network connection methods in LintClient.

This CL also fixes this bug:
62027560: Dependencies added with older versions while adding widgets

I initially thought this bug should be fixed by the new repository,
e.g.  it was using the old versions because it couldn't "see" the
maven.google.com versions until the above fix. However, we shouldn't
just blindly use the latest available version either. When you drag in
a recyclerview in a project, we should use the same version of
recyclerview as the existing support lib version used in the IDE.
That's what this CL fixes: it now looks up the existing version of
appcompat, if any, and if found uses that to resolve dynamic versions
in support library version resolution.

Test: Most of this is tested in tools/base; the library version
resolution code is tested by the project instantiation tests.

Change-Id: I02c55ebc98cde66aa78d4c9facf02c7d34af8e46