Update TemplateTest to use form factor appropriate modules
This CL updates TemplateTest to create an Application
module that is of the same form-factor as the activity it
is testing. This extends TemplateTest to cover all the
new module templates, and allows Wear templates to be tested

Note that AndroidAuto isn't a real form factor in the
same way the others are, so we still need to use a Mobile
module for android auto activities.

Fixes: http://b/37703639
Test: TemplateTest
Change-Id: I106d71e6ee06eecd7b681d80c293cafa9be49b0c