Now *new* sync uses embedded repo for new projects.
The purpose of the embedded repo is to improve user
experience for first-time users by allowing them to
create new projects without requiring an Internet
connection. Subsequent Gradle syncs should not need
the embedded repo really.

This CL also cleans up CommandLineArgs, by moving creation of
the init script that applies the Java library plugin to the
class GradleInitScripts.

Bug: 38463221

(Partially fixes the bug. The final fix will be to do the
same for build execution.)

Test: JUnit tests.
Change-Id: I937b722145f5ce52b06ac1b54cf666b91b7bba41