Allow components to be moved up and left
This introduces the concept of a view offset, which allows there to be a
view-specified translation between android coordinate space and swing
coordinate space separate from the positioning of the view itself in the
design surface.
In practice, this allows us to place scene components at an arbitrary
place (highly positive or negative) in the android system, and keep the
view sized and positioned sanely.

TODO: limit the rate at which the view expands/contracts. Currently
dragging something out of or especially back toward the bounding region
of other coordinates causes the view to expand or shrink too quickly to
be useful.

Test: Updated NavSceneTest
Change-Id: I0d1777fdc8d8b4dd3f219c3f352f84af575e04f7