Fixes erasing the formating when changing a theme
A Component constructed by ColumnTreeBuilder was erasing the formating
when changing the theme(e.g from Darcula or to Darcula).
For example, ColumnTreeBuilder was used by CPU profiler and memory profiler.

The root issue was that when changing the theme IntellIJ traverses
all components and calls updateUI() from them, in a result it will override
the custom ui (i.e ColumnTreeUI), which is needed to adjust width of rows.

Bug: 35952312, 62478709
Test: ColumnTreeBuilderTest#updatingUIShouldPreserveColumnTreeUI
Change-Id: Ie88ed0944077d8f9ed9ccd1d572e6a5252c2c190