Fix NewProjectTest.testCreateNewMobileProject
The module name is not the exactly same to the project name if the
same project name has been used. E.g., the module name will be
"Test Application2" if there is already a "Test Application" project
and the project name is used again.

Because "Test Application" has been used more than once in sanity
tests, this test must have a unique project name in order to pass
the verification that the module name is the exactly same to the
project name.

Test: run the test.
Change-Id: Ic7fa56a293b39faa15f43e63c27859ca9be6958a