Only track device info with CHANGE_DEVICE event
Previously, we would send the user's target device with every profiler
event, just in case we needed to correlate it, and since the deviceinfo
messsage was small. Now, however, we are tracking more information
(device manufacturer and model), which would be unecessarily expensive
to send out with each and every profiler event. So, we change the logic
to only send out the information when we know it changes.

Bug: 62802911
Test: N/A
Change-Id: I8e6e932a3d71632325422c5ce6055d641d76d8f7