[Bug 37138905] Remove BuildOutputViewCleanUpTask from PreSyncProjectCleanUp
Bug 37138905: Gradle console output is cleared on Build APK because
PreSyncProjectCleanUp calls to cleanUp of BuildOutputViewCleanUpTask.

The purpouse of doing that is to have a clean console everytime sync starts,
but that is already done by GradleTasksExecutorImpl.

This change removes BuildOutputViewCleanUpTask from PreSyncProjectCleanUp

Test: Adapted tests to not use this task and manually confirm that console
output is still present after a build.

Change-Id: I8fced79e65fcd0736e75aa81de45e91957be43de