[AIA] Make Studio skip provision if already been provisioned
Now Studio keeps a set of provisioned devices in the provision task
and queries it before provisioning to not do the work again when not
needed.

This CL adds an UI so the user can choose to erase that cache. It also
automatically erases the cache after 1 day.

Test: manual

Bug b/37684779

Change-Id: Ie1c2956c580887aad335c46f6b6b8630f9765a12