Ignore testRunner written in workspace.xml when needed
Instrumentation test runner in android test run configurations' xml was
being considered when not null but hidden from the run config editor.
The right behavior is to consider it (and make the editor visible) only
for non gradle projects. For gradle projects the one in the DSL should be
considered.
This CL fixes it.

Test: AndroidTestRunnerTest#testRunnerComponentsEmptyForGradleProjects

Bug b/62791331

Change-Id: I1f5e2fa83da849387ac3e23191f45267f4f79800