[AIA] Fix first run of Instant App for N emulators
Because of b/63257876, first (provision + run) of an instant app for N
emulators was not working: the same instance of the IDevice is used
during all the process, but in first provision, when getting root access
to the emulator, the connection fails that instance becomes offline,
what causes problem when trying to read from the logcat while side
loading.

This CL fixes the problem by trying to get an online instance of IDevice
of the same device before side loading the app, if the current instance
of IDevice is not online.

Test: manual

Bug b/63163062

Change-Id: Ic6e60da5548216d6179a368baa24009a93c8990a