Only show possible resizing options in LinearLayout
If a widget has aligned to some edge of parent, it
should not show the resizing options.
For example, if a widget has left and top gravity in
the LinearLayout, it should not have resizing options
on top and left side since it cannot increase/decrease
the width and/or height from those sides.

Test: Passes all tests in LinearLayoutHandlerTest
Bug: 63562276

Change-Id: I9c8fc9096b2520ac028138f5f7b15a18b9164771