Move ImageUtils.loadIcon into ShadowPainter
The loadIcon method was used to load an icon from the android plugin.
We'd like to move ImageUtil into adt-ui to keep it independent of the
android plugin, and this is the only method in it that assumes the
android plugin. Since it was used in just a single class, the method has
been moved into that class.

Bug: 63768825
Test: existing tests pass
Change-Id: I73c91ef119690119d246099ccdbbd90eab04eff5