[Bug: 63534751] Handle errors when Android Gradle plugin cannot be found
This CL adds a new sync error handler to process messages related to not
finding the Android gradle plugin. The error handler generates quickfixes
following these rules:

- If Google Maven repository is not in the repositories of the buildscript
block of the build.gradle file where the plugin version iss defined:
  - Quick fix to add repository
  - Hyperlink to open build.gradle

- If Google repository is already in the
repositories:
  - Hyperlink to open build.gradle.

The quick fix adds "google()" if Gradle version is at least 4.0, if not, it
adds this:

maven {
  url "https://maven.google.com/"
  name "Google"
}

Test: Add new tests for new error handler, new hyperlink, tests for
handling google repository in BuildScriptModelTest and
RepositoriesModelTest, and a test for isGradle4OrNewer.

Screenshots:
- Add and open hyperlink: https://screenshot.googleplex.com/oPGb15D4FoB
- Open hyperlink only: https://screenshot.googleplex.com/M9C0DmSZV91

Bug: 63534751

Change-Id: I08b7a8a96c54137eb25e733fa1556bc475e45ab6