Add dialog to warn users about large trace files
Parsing large CPU trace files (100MB+) is expensive and can consume a
lot of both memory and time. However, some users might have powerful
machines and don't bother with the trace file. Minding these users and
the ones with regular machines, which probably want to avoid parsing
large trace files, a dialog was created when a trace file is too large.
The users have the option to abort parsing before it starts or proceed.

Test: Unit tests + manual
Bug: 38389851
Change-Id: I83fd8882b76858d549d6e3d83fda9c2f5fd8e2cc