Provide auto-completion for include tag in layout
Auto-completion for android:layout_[attr_name] attributes
should be enable in <include> tag.

For example, <include> tag in RelativeLayout should
provide the auto-completion of android:layout_alignXXX
attributes.

This also works for customized domain name. For instance,
the auto-completion of app:layout_constraintXXX attributes
in Constraintlayout is supported as well.

Test: Added and passed the test cases in AndroidLayoutDomTest
Bug: 63797631
Change-Id: Idd582cca37c8c6d3a65bdb3d6a76ff1aa6cacc37