Moving common parts of uibuilder to "common"
This almost universally involves only moving files (some tests' tearDown
was added or enhanced to null out fields, since I got OOM when running
all the designer tests on my machine).

This should make it easier to keep track of what code is expected to be
part of the "framework" and what should be specific to the layout/nav
editors.

It also exposes some gaps in the framework-level testing...

Note that this doesn't include any of the properties editing framework.
That seems more complicated and will come in a separate CL.

Test: just moving things between packages, no additional tests needed.

Change-Id: Ia947312dd4390b6ff1c04652883ce64710c0983b