Context click on "manifests" to create activity produces AssertionError
If we are showing the Android Perspective, some of the "folders" shown
are just a representation, and do not map to real folders. For example
"Manifests" shows all Application manifests and there could be more than
one in different directories. For this "Simulated" folders, the "Action
Target Directory" is null. This CL allows null to be a valid value.

Bug: 38174948
Test: N/A See bug for details
Change-Id: Iecbb16d43cd2c0d320b1f5ff436cc9bcea4e1f67