Fix CpuCaptureStage leaking due to dependency not removed
AspectModel#removeDependencies was removing the dependencies of
CpuProfilerStage from the observer itself, but was keeping all of them (but
one) in the models it listens to. As a result, when these models
changed, we would trigger the methods from CpuProfilerStage.

This CL fixes the root cause in AspectModel, add a unit test to
AspectModelTest to cover it, and adjust the tests of CpuProfilerStageTest
that would be flaky if the issue was not fixed.

Bug: 64139004
Bug: 64203834
Test: Unit tests
Change-Id: I965507ea6883d5774532614ba95fb088b01c4e71