Obtain jar in PublicMethodsBuildActionTest through class loader
Following CL ag/2630850, this CL changes the way we obtain the Guava jar
to be tested: instead of just using the current version (19.0) we now
get the used jar from the code source (or resources) of a class
contained in the jar.

This CL also adds a test for making sure that BuildActions contained
Guava in public methods would fail the test.

Test: PublicMethodsBuildActionTest

Bug b/63929984

Change-Id: Ib617063812e4508a3375177c9f68bb512115ca4c