Close project when test is finished to prevent FileNotFoundExceptions
Android Studio saves data to a .iml___jb_tmp___ file in the APK
project's directory when the project is closed. Projects are normally
closed automatically by the test without issues, but
CreateAPKProjectTest will recursively delete the APK projects
directory in the @After method. The automatic closing of projects
is done after the @After method is called, leading to a
FileNotFoundException when the IDE tries saving to the file.

This change prevents the exception from being thrown by first
closing the project before deleting the ApkProjects directory.

Test: Run this test
Change-Id: I84c4da16d8f05a9bbdde6fe97cee5fb00af9224e