Don't try to stop profilers when it's already stopped
Calling StudioProfilers#stop() will stop the updater, set the device to
null and fire ProfilerAspect.STAGE aspect. This only makes sense when
the profilers are running. If they are already stopped, it's not the
responsibility of stop() method to to set the device to null and fire
the aspect.

That was causing our models to be inconsistent. For instance:
  * AndroidProfilerToolWindow listens to ProfilerAspect.STAGE and calls
    updateToolWindow() if this aspect is fired.
  * updateToolWindow checks if the profiler is stopped and, if it is,
    the method will indirectly call AndroidProfilerToolWindow#dispose()
  * dispose() calls StudioProfilers#stop(), which was again will fire
    ProfilerAspect.STAGE.

Bug: 63745419
Test: added unit test
Change-Id: Ieae24c26037077ab6b06f6a32d3ba4a7c74a506a