Get ResourceRepositories inside read actions
Getting a ResourceRepository used to acquire a specific lock
dependending on the type of repository, and then potentially a read
lock.
But in some places, those get*Resources methods are called inside a read
lock, which can cause a deadlock. So we make sure to always be in a read
action before locking the specific repository.

Bug: 64072963
Test: NlEditorTest.testAddDesignLibrary regularly deadlocked on buildbot
Change-Id: Ibfb158a4efe94c4b0cc2f49adfa8787a02a86de1