SimpleperfTraceParser should add a node with thread name
We were adding a special node to represent the root of the call stack
corresponding to a thread. However, we were always naming it "main". We
should be consistent with other parsers (e.g. VmTraceParser) and use the
thread name instead, falling back to "main" only when a name is not
available.

Test: SimplePerfTraceParserTest
Bug: 37294023
Change-Id: I05e1a1f4c6c6eb0a3c8b7bcf7bbd6f271d92afe9