Stop compiling test sources when not needed
This CL fixes two problems:
- First, since TestCompileType.NONE was returning all the available test
artifacts in a variant, and in this specific case it should return no
artifact. This CL renames NONE to ALL and introduces NONE.
- Second, when running test (android tests or unit tests) in a module
depending on others, the test source for all the modules were compiled
and not only for the running module.

Test: GradleModuleTasksProviderTest
Fix bug b/37756694

Change-Id: Ib3af76889e454af6c6e02cf6cf37d0526f3cbc6b