[APK Debugging] Added APIs to mark a project as "APK Debugging" project.
Before this CL, when reopening an "APK Debugging" project, a
"Gradle-aware Make" was added to its run configuration. The reason
is that when the task is created, the check
AndroidProjectInfo#isApkProject return false because modules with
the ApkFacet were not loaded yet.

The fix is to mark the project as "APK Debugging" project, in addition
to the ApkFacet in modules.

This CL adds the APIs to mark a project, but the actual "marking"
happens during project setup, in a different CL.

Bug: 64766060

Test: JUnit tests.
Change-Id: I769d5179204f4aad87d7d11ff13d13fc178d589e