[AVD Mgr] Modify the "Cold Boot Now" behavior
The "Cold Boot Now" behavior was the same as if the general "Cold
Boot" option was selected. This option does not save a snapshot
on exit, so the subsequent launch would not be Quick.

For "Cold Boot Now," give a different option to the Emulator so it
does a cold boot, but saves its state on exit so the next invocation
can be Quick.

(The general cold boot option is "-no-snapstorage"; the option for
Cold Boot Now is "-no-snapshot-load".)

Bug: 64851650

Test: Expanded AvdManagerConnectionTest::testAddParameters()

Change-Id: I2b709042110f017f6f69efa61b9b50af5f0949b9
(cherry picked from commit dc435863fcb4fcacbf268709b124724b816430ea)