Update BuildSystemService sync API
This CL expands the API provided by BuildSystemService#syncProject
so that existing usages of GradleSyncInvoker#requestProjectSync
can be migrated to BuildSystemService, which is build system
agnostic.

This CL also updates the Gradle implementation of BuildSystemService
to support the API expansion and updates existing usages of the
syncProject method to reflect changes to the method signature.

Bug: 64436699
Test: minor changes to existing tests
Change-Id: Iad7b5093ada7eae488c4970471de90c637cd6d04