Device File Explorer: improve support for userdebug devices
When pushing/pulling file to/from system directories on userdebug
devices, adb returns a TRANSFER_PROTOCOL_ERROR for permission
denied error (as opposed to NO_REMOTE_OBJECT on user devices).

We need to detect this additional error condition to detect we
need to fallback to using the "push to temp, cp temp destination"
combination of commands.

Most of the code in this change is for unit testing, the only actual
change is in AdbDeviceDirectFileEntry.

Note that this code does not "fix" bug 65165623, as it is not clear
this is an actual bug, but this change is anyways related, as this
we observed on a userdebug device where the build.prop file had
restricted read access.

Bug: 65165623
Test: Add specific tests
Change-Id: Ic9ebe86bb447e5771f75cf14da4948d11d852066
(cherry picked from commit 7fa547fbd2138782f93da797469faa82f216a24d)