Populate javadoc paths for dependencies.
This CL fixes the bug when the paths of type DOCUMENTATION were
never populated and hence the javadoc roots were never attached
to the libraries.

Now both when AAR and JAR dependencies are constructed, the
javadoc lookup will be performed and the paths populated within
the dependency representation accordingly.

This in turn will make gradle sync add the corresponding root
types to the library, which has the corresponding calls in place
but they were a no-op because the documentation paths were not
populated in the first place.

Bug: b/37125917
Test: Added LibraryDependency.testConstructorWithAarAndJavadoc and
adjusted Android module setup tests to make sure javadoc roots are
set when populated in the dependency, and that the lookup itself
is not performed during setup.

Change-Id: I682cb58a5231f36e3ecd85c3869708740477f034