CPU capture shouldn't be set to null after entering L3
We were setting the capture back to null and closing the capture details
panel when the user selected an area that doesn't include any captures.

Once the user enters L3 and the capture details panel is open, we
shouldn't close it until the user explicitly goes back to L1. When they
select an area with no captures, we should keep the selection of the
last capture.

Change-Id: I78b4676d56a47d0aade3539ea34ef7165223fa19
Fixes: 65158869
Test: Unit test + manual