Add default project system tests and new test rules
Created AndroidProjectRule as a way to access android projects and modules
in a JUnit 4-style test case, and a way to perform mocking that always
cleans up leaked mocks. Most existing test cases do not remove their
mocks, facets, or overridden services from the project in their
teardown methods. This rule guarantees that such cleanup always occurs
without any requirement for the test cases to do anything in
their teardown.

It uses a shared Project instance, which is more lightweight and faster
to construct. However, any modifications to the Project should be done
via the Rule to ensure they are properly cleaned up on test teardown.

Created NamedExternalResource as a way to expose the test name to
JUnit4-style test rules.

Test: DefaultProjectSystemTest
Change-Id: I5e589b558851a4cfe010af4f8d90d88268884919