Fix selection bug in new palette when typing in search box.
This CL is fixing a bug in the new palette where typing in
the search filter causes changes to the category list.
The code is attempting to keep up the currently selected
category but this fails with an IndexOutOfBoundsException
because the selection model has not been updated yet.

The solution here avoids calling list.getSelectedValue
during filter value changes. Instead we rely on the last
recorded selected value we get from the selection model.

Bug: b/64812850

Test: Added ui tests to verify that this works.

Change-Id: I733670ff40f7c6a7ff1d3e83b2d4ae53d88d42f5