Fixed exceptions in live alloc when unplug + replug
There are two problematic cases this CL fixes -
1. A change was made during refactoring in bug 64490742, which we
pass in a range(Long.MIN_VALUE, Long.MAX_VALUE) to determine if we
already have an ongoing AllocationsInfo. The range is converted from
us to ns subsequently which causes overflows, in which case no valid
AllocationsInfo can be found when in fact there is one.
2. We need to guard against the resume/suspend allocation tracking
calls, as the grpc channel from perfa->perfd may not be available at
that time (e.g. when a device is plugged in, the channel is estebalished
via AttachAgent which happens after the resume logic)

Bug: 65484857
Test: Updated unit tests
Change-Id: I6a18f4b8182a2be7a0b2b6b2fc382dc547d3d558
(cherry picked from commit a33c18630e3be966c646f1949bfaca7ca41db356)