Adds watchface template test to KNOWN_BROKEN.
Current support library for wearbles (2.0.4) breaks with watch faces.
The next version will be released (hopefully) on Monday. After
that has been released, we can take this test out of KNOWN_BROKEN.

We have some watch face changes going in as well, but they have been
tested throughly on 2.0.3 and 2.0.5.

Bug: 65062154
Test: Manually tested.
Change-Id: I02127414e0f508ca8f17c3866be5f8f2f7d9626e