Use ViewHandler to browse for classes for <view> and <fragment>.
The browse button in the property inspector & property table was
calling ChooseClassDialog.openDialog directly. This caused us to
have duplicated logic in the ViewHandlers and the code for the
Browse button.

Remove this code duplication by using the ViewHandler to bring up
the ChooseClassDialog. Add a new method to the ViewHandler API for
getting the browser action.

Test: Updated
Change-Id: Id91a5dc12b3472f5e24824818e6f7218705d6712