Adds Kotlin watchface unit test.
Details: Adds Kotlin specific unit test for wear watch faces.
Because the Kotlin test just uses WatchFaceService (which is already
listed on KNOWN_BROKEN), I do not need to include it again.

More info on why it is broken:
Current support library for wearbles (2.0.4) breaks with watch faces.
The next version will be released (hopefully) on Monday. After
that has been released, we can take this test out of KNOWN_BROKEN.

We have some watch face changes going in as well, but they have been
tested throughly on 2.0.3 and 2.0.5.

Bug: 65062154
Test: Manually tested, also this is the unit test.
Change-Id: Iaf50b99840f96238a55af87ced7eb6273aee689b