[AIA] Fix manifest merge for feature modules
When merging the manifest, if the project type is "feature", its
manifest should be merged as an application merge (libraries manifests
should be merged in it) and not as a library merge.

This CL guarantees the manifest is merged correctly for features.

Test: FeatureMergedManifestTest
Fixes bug 65366849

Change-Id: I2fdfdd692bd22308ac61f9c4b95711c77ac3f64e
(cherry picked from commit d93b53a17493dab83ab6eebde14b01839644be51)