Replace GradleSyncState usage in ManifestInfo
Currently, ManifestInfo$ManifestFile uses GradleSyncState to get
the timestamp of the latest sync for a project, which creates an
unnecessary dependency on gradle.

This CL removes this dependency by having ManifestFile use the
build system independent sync listening API provided by
AndroidProjectSystem to locally keep track of the latest sync time
for its project.

This CL also updates DefaultProjectSystem#syncProject so that it
broadcasts the result of its immediate failure for test purposes.

Bug: 65626649
Test: ManifestInfoTest
Change-Id: I3e29c03fcf6bc60878dd46bdfe7986b1c3375950